requirements.txt
django==4.2.1
flake8==6.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.5
pep8-naming==0.13.3
django-debug-toolbar==4.0.0
django-crispy-forms==2.0
asgiref==3.6.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
